Server Restructure (Copilot Assisted)
