+++ .env.example
@@ -1,0 +1,14 @@
+# Environment Configuration
+# Copy this file to .env and update values as needed
+
+# Server port
+PORT=3000
+
+# Node environment (development, production)
+NODE_ENV=development
+
+# Session secret (change this in production!)
+SESSION_SECRET=interlinked-game-session-secret-key-2025
+
+# Database path (relative to project root)
+DB_PATH=database/interlinked.db
