+++ .gitignore
@@ -1,8 +1,34 @@
+# Database
 database/
-.git/
+
+# Dependencies
+node_modules/
+
+# Environment
+.env
+.env.local
+.env.*.local
+
+# Logs
 *.log
+npm-debug.log*
+
+# Temp files
 *.tmp
 **/Temp/
 **/temp/
 *.bak
-node_modules/
+
+# Editor
+.vscode/
+*.swp
+*.swo
+.idea/
+
+# OS
+.DS_Store
+Thumbs.db
+
+# Build outputs
+dist/
+build/
