+++ package.json
@@ -1,19 +1,29 @@
 {
-  "name": "new",
+  "name": "interlinked-website",
   "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
+  "description": "Interlinked Creations website with chat, friends, and online features",
+  "main": "src/server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node src/server.js",
+    "dev": "node --watch src/server.js",
+    "db:setup": "node scripts/dbSetup.js",
+    "db:update": "node scripts/dbUpdate.js",
+    "db:query": "node scripts/dbQuery.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [],
-  "author": "",
+  "keywords": [
+    "interlinked",
+    "chat",
+    "websocket",
+    "express"
+  ],
+  "author": "Interlinked Creations",
   "license": "ISC",
   "type": "commonjs",
   "dependencies": {
     "bcrypt": "^6.0.0",
     "better-sqlite3": "^12.2.0",
+    "dotenv": "^16.6.1",
     "express": "^5.1.0",
     "express-session": "^1.18.2",
     "ws": "^8.18.3"
