+++ scripts/dbQuery.js
@@ -1,20 +1,26 @@
+/**
+ * Database Query Script
+ * Run ad-hoc queries against the database
+ * 
+ * Usage: node scripts/dbQuery.js
+ */
 const Database = require('better-sqlite3');
 const fs = require('fs');
 const path = require('path');
 
 // Database configuration
-const DB_PATH = path.join(__dirname, 'database', 'interlinked.db');
-const DB_DIR = path.join(__dirname, 'database');
+const DB_PATH = path.resolve(__dirname, '../database/interlinked.db');
+const DB_DIR = path.dirname(DB_PATH);
 
 // Ensure database directory exists
 if (!fs.existsSync(DB_DIR)) {
     fs.mkdirSync(DB_DIR, { recursive: true });
 }
 
-// Create new database connection
+// Create database connection
 const db = new Database(DB_PATH, { verbose: console.log });
 
-// Run the query
+// Run the query - modify this as needed
 const DBQuery = `
     SELECT *
     FROM Users
