+++ scripts/dbSetup.js
@@ -1,0 +1,7 @@
+/**
+ * Database Setup Script
+ * Run this to create a fresh database with all tables
+ * 
+ * Usage: node scripts/dbSetup.js
+ */
+require('../src/db/setup');
